internal/auth: fix circular dep
