+++ service/auth/server/rules/rules.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/micro/go-micro/v3/auth"
 	gostore "github.com/micro/go-micro/v3/store"
-	"github.com/micro/micro/v3/internal/namespace"
+	"github.com/micro/micro/v3/internal/auth/namespace"
 	pb "github.com/micro/micro/v3/proto/auth"
 	"github.com/micro/micro/v3/service/errors"
 	"github.com/micro/micro/v3/service/logger"
