+++ service/runtime/server/handler.go
@@ -6,7 +6,7 @@ import (
 
 	goevents "github.com/micro/go-micro/v3/events"
 	gorun "github.com/micro/go-micro/v3/runtime"
-	"github.com/micro/micro/v3/internal/namespace"
+	"github.com/micro/micro/v3/internal/auth/namespace"
 	pb "github.com/micro/micro/v3/proto/runtime"
 	"github.com/micro/micro/v3/service/auth"
 	"github.com/micro/micro/v3/service/errors"
