+++ service/store/server/handler.go
@@ -10,7 +10,7 @@ import (
 	"time"
 
 	gostore "github.com/micro/go-micro/v3/store"
-	"github.com/micro/micro/v3/internal/namespace"
+	"github.com/micro/micro/v3/internal/auth/namespace"
 	pb "github.com/micro/micro/v3/proto/store"
 	"github.com/micro/micro/v3/service/errors"
 	"github.com/micro/micro/v3/service/store"
